requirements.txt
pandas>=1.5.0
requests>=2.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
numpy<2.0
dspy-ai>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
